// src/lib/api/generated/models/MinecraftPlayerProfile.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PolystirolHub Backend
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { CurrentNickname } from './CurrentNickname';
import {
	CurrentNicknameFromJSON,
	CurrentNicknameFromJSONTyped,
	CurrentNicknameToJSON,
} from './CurrentNickname';
import type { LastSeen } from './LastSeen';
import { LastSeenFromJSON, LastSeenFromJSONTyped, LastSeenToJSON } from './LastSeen';
import type { Platform } from './Platform';
import { PlatformFromJSON, PlatformFromJSONTyped, PlatformToJSON } from './Platform';

/**
 * Профиль игрока
 * @export
 * @interface MinecraftPlayerProfile
 */
export interface MinecraftPlayerProfile {
	/**
	 *
	 * @type {any}
	 * @memberof MinecraftPlayerProfile
	 */
	uuid: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof MinecraftPlayerProfile
	 */
	name: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof MinecraftPlayerProfile
	 */
	registered: any | null;
	/**
	 *
	 * @type {CurrentNickname}
	 * @memberof MinecraftPlayerProfile
	 */
	currentNickname?: CurrentNickname;
	/**
	 *
	 * @type {Platform}
	 * @memberof MinecraftPlayerProfile
	 */
	platform?: Platform;
	/**
	 *
	 * @type {LastSeen}
	 * @memberof MinecraftPlayerProfile
	 */
	lastSeen?: LastSeen;
	/**
	 *
	 * @type {any}
	 * @memberof MinecraftPlayerProfile
	 */
	totalPlaytime: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof MinecraftPlayerProfile
	 */
	totalKills: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof MinecraftPlayerProfile
	 */
	totalDeaths: any | null;
	/**
	 *
	 * @type {any}
	 * @memberof MinecraftPlayerProfile
	 */
	serversPlayed: any | null;
}

/**
 * Check if a given object implements the MinecraftPlayerProfile interface.
 */
export function instanceOfMinecraftPlayerProfile(value: object): boolean {
	let isInstance = true;
	isInstance = isInstance && 'uuid' in value;
	isInstance = isInstance && 'name' in value;
	isInstance = isInstance && 'registered' in value;
	isInstance = isInstance && 'totalPlaytime' in value;
	isInstance = isInstance && 'totalKills' in value;
	isInstance = isInstance && 'totalDeaths' in value;
	isInstance = isInstance && 'serversPlayed' in value;

	return isInstance;
}

export function MinecraftPlayerProfileFromJSON(json: any): MinecraftPlayerProfile {
	return MinecraftPlayerProfileFromJSONTyped(json, false);
}

export function MinecraftPlayerProfileFromJSONTyped(
	json: any,
	ignoreDiscriminator: boolean
): MinecraftPlayerProfile {
	if (json === undefined || json === null) {
		return json;
	}
	return {
		uuid: json['uuid'],
		name: json['name'],
		registered: json['registered'],
		currentNickname: !exists(json, 'current_nickname')
			? undefined
			: CurrentNicknameFromJSON(json['current_nickname']),
		platform: !exists(json, 'platform') ? undefined : PlatformFromJSON(json['platform']),
		lastSeen: !exists(json, 'last_seen') ? undefined : LastSeenFromJSON(json['last_seen']),
		totalPlaytime: json['total_playtime'],
		totalKills: json['total_kills'],
		totalDeaths: json['total_deaths'],
		serversPlayed: json['servers_played'],
	};
}

export function MinecraftPlayerProfileToJSON(value?: MinecraftPlayerProfile | null): any {
	if (value === undefined) {
		return undefined;
	}
	if (value === null) {
		return null;
	}
	return {
		uuid: value.uuid,
		name: value.name,
		registered: value.registered,
		current_nickname: CurrentNicknameToJSON(value.currentNickname),
		platform: PlatformToJSON(value.platform),
		last_seen: LastSeenToJSON(value.lastSeen),
		total_playtime: value.totalPlaytime,
		total_kills: value.totalKills,
		total_deaths: value.totalDeaths,
		servers_played: value.serversPlayed,
	};
}